Revert README.txt to its original name
The new file was not being included in the tarball, since it was
forgotten to add this file to the needed Makefile.am. This could be
fixed, but it seemed simpler to use the old filename, which will be
included automatically.

Plus, the old name of README is standard GNU and matches the convention
of some of the other documentation files.