Remove PDF documentation from Debian package
The PDF files are included when building a source and binary package at
the same time, but not included if the binary package is built from the
source package.

An expert in CDBS could fix this, but since the Debian project has teckit
without CDBS in experimental, I am not going to take the time to fix the
current CDBS packaging.