Update to Unicode 15.0.0
The update script was changed to use curl instead of wget.
The wget command was being run with the timestamp option (-N)
which would fail if the local repo had been created after
the files had been updated at unicode.org (therefore having a newer
timestamp). Turning off the timestamp option would cause wget
to create a file with a numeric suffix (NormalizationTest.txt.1)
which would not be read for running tests.